+++ qmi/instruments/yokogawa/__init__.py
@@ -1,0 +1,7 @@
+"""
+Yokogawa, oscilloscopes.
+
+The qmi.instruments.yokogawa package provides support for:
+  - DLM4038 model.
+"""
+from qmi.instruments.yokogawa.dlm4038 import Yokogawa_DLM4038 as Yokogawa_Dlm4038