+++ CHANGELOG.md
@@ -8,13 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Thorlabs KDC101 controller QMI driver. It can at the moment control Z906, Z912 and Z925 actuators and PRMTZ8 rotation stage.
-- New QMI driver for Agiltron FF1x8 optical switch
+- New QMI driver for Agiltron FF1x8 optical switch.
+- New QMI driver for Yokogawa DLM4038 oscilloscope.
 
 ### Changed
 - The Agiltron FF optical switch QMI drivers have now common base class in `qmi.instruments.agiltron._ff_optical_switch`
 - Refactored also unit-tests for Agiltron FF optical switches.
 
 ### Fixed
+- In `usbtmc.py` now doing `.strip()` on `dev.serial_number` string to avoid SNs with whitespace character(s).
 
 ### Removed